fix: ajustes gerais

+++ src/products/services/deleteProductService.ts
@@ -1,10 +1,14 @@
-import { Product } from '@prisma/client';
+import { Response } from 'express';
 import { ProductRepository } from '../repositories/productRepository';
 
 export class DeleteProductService {
   constructor(private productRepository: ProductRepository) {}
 
-  async deleteProduct(id: string): Promise<Product | null> {
-    return this.productRepository.delete(id);
+  async deleteProduct(id: string, response: Response) {
+    const verifyProduct = await this.productRepository.findById(id);
+    if (!verifyProduct) {
+      return response.status(404).json({ message: 'Produto não encontrado' });
+    }
+    await this.productRepository.delete(id);
   }
 }
